test(guard): isolate network status client coverage
